requirements.txt
requests==2.32.3
python-dotenv==1.0.1
google-api-python-client==2.169.0
google-auth==2.40.3
